# Move exponential_ from TH to Aten (CPU)