Summary:
This diff will address https://github.com/pytorch/pytorch/issues/24699

We ask the input `lambda` to be >= 0 to be same as https://docs.scipy.org/doc/numpy-1.15.0/reference/generated/numpy.random.exponential.html#numpy-random-exponential. This does not exist in the previous implementation.

Benchmark script
```
import torch
import torch.nn as nn
import time

torch.manual_seed(0)

def _time():
    return time.time()

device = "cpu"

#warm up
for n in [10, 100, 1000]:
    input = torch.randn(128, n, requires_grad=False, device=device)
    for i in range(1000):
        input.exponential_()

for n in [1, 10, 100, 1000]:
    fwd_t = 0
    input = torch.randn(128, n, requires_grad=False, device=device)
    for i in range(10000):
        t1 = _time()
        input.exponential_()
        t2 = _time()
        fwd_t = fwd_t + (t2 -t1)
    fwd_avg = fwd_t / 10000 * 1000
    print("input size(128, %d) forward time is %.4f (ms)." % (n, fwd_avg))
```
Output
```
================================================================================
Before the change, Program Output:
================================================================================
Higher priority item already registered, skipping registration of LINUX
input size(128, 1) forward time is 0.0129 (ms).
input size(128, 10) forward time is 0.1069 (ms).
input size(128, 100) forward time is 1.0409 (ms).
input size(128, 1000) forward time is 10.3827 (ms).
================================================================================
After the change, Program Output:
================================================================================
2020-01-22 10:52:06,406 fbpkg.fetch INFO: completed download of bento_kernel_pytorch:641
2020-01-22 10:52:06,407 fbpkg.fetch INFO: extracted bento_kernel_pytorch:641 to bento_kernel_pytorch
All packages downloaded successfully
Higher priority item already registered, skipping registration of LINUX
input size(128, 1) forward time is 0.0129 (ms).
input size(128, 10) forward time is 0.1065 (ms).
input size(128, 100) forward time is 1.0374 (ms).
input size(128, 1000) forward time is 10.3569 (ms).
================================================================================
```

Test Plan: Sandcastle and Github tests

Differential Revision: D19518700