@kostmo 2020-01-22
## :pill: CircleCI build failures summary and remediations
As of commit fbbd55cc:


* **1/1** failures introduced in this PR


## Detailed failure analysis

One may explore the probable reasons each build failed interactively [on the Dr. CI website](https://dr.pytorch.org/commit-details.html?sha1=fbbd55cc352407b3d04422ba3a59502ca2e9c5a8).

### :detective: 1 new failure recognized by patterns
The following build failures do not appear to be due to upstream breakage:
#### [![See CircleCI build](https://avatars0.githubusercontent.com/ml/7?s=12)](https://circleci.com/gh/pytorch/pytorch/4319144) pytorch_macos_10_13_py3_test (1/1)
**Step:** "Test" ([full log](https://dr.pytorch.org/api/view-log-full?build_id=89550513) | [pattern match details](https://dr.pytorch.org/build-details.html?build_id=89550513))
<details>
<summary>
<code>Jan 22 14:11:41 RuntimeError: test_autograd failed!</code>
</summary>

```
Jan 22 14:11:40 Ran 1046 tests in 283.128s 
Jan 22 14:11:40  
Jan 22 14:11:40 FAILED (errors=1, skipped=13, expected failures=1) 
Jan 22 14:11:40  
Jan 22 14:11:40 Generating XML reports... 
Jan 22 14:11:41 Traceback (most recent call last): 
Jan 22 14:11:41   File "test/run_test.py", line 457, in <module> 
Jan 22 14:11:41     main() 
Jan 22 14:11:41   File "test/run_test.py", line 450, in main 
Jan 22 14:11:41     raise RuntimeError(message) 
Jan 22 14:11:41 RuntimeError: test_autograd failed! 
Jan 22 14:11:41 + cleanup 
Jan 22 14:11:41 + retcode=1 
Jan 22 14:11:41 + set +x 
```

</details>



---

<details><summary>This comment was automatically generated by <a href="https://github.com/kostmo/circleci-failure-tracker/tree/master/docs/from-pull-request-comment">Dr. CI</a> (expand for details).</summary>Follow <a href="https://dr.pytorch.org/admin/comments-opt-out.html">this link to opt-out</a> of these comments for your Pull Requests.

Please report bugs/suggestions on the <a href="https://github.com/kostmo/circleci-failure-tracker/issues">GitHub issue tracker</a>.
</details>