@hl475 2020-01-27
address comments from @VitalyFedyunin