@hl475 2020-01-27
add more test coverage and explicitly comment for the case when lambda is 0.