@zou3519 2020-04-16
The case where `lambda=0` should be different from `scale=0` in the numpy case, right?
numpy uses the `scale` parameter, which is equivalent to `1/lambda` in our API https://docs.scipy.org/doc/numpy-1.15.0/reference/generated/numpy.random.exponential.html#numpy-random-exponential  https://pytorch.org/docs/stable/tensors.html#torch.Tensor.exponential_

This PR changed the behavior of `lambda=0` to return a tensor full of zeros, but it should really match the behavior of `numpy.random.exponential(scale=99999999, shape=10)` (some big number) and return a bit number.

Does that sound right,  @hl475 @VitalyFedyunin ?