@gchanan 2020-04-17
I think both scale and the result should be `inf`, i.e.:

```
>>> np.random.exponential(np.array(1.)/np.array(0.))
//anaconda2/envs/pytorch1.4/bin/ipython:1: RuntimeWarning: divide by zero encountered in true_divide
  #!//anaconda2/envs/pytorch1.4/bin/python
inf
```